Version 1.2
Lots of changes from Romain Chossart, thanks for that!

- The buffer is not displayed (actually, it is displayed then deleted
quickly) when "no Python doc found". If the buffer was already
displayed, it does a "u" to go back to the last pydoc. A message is
displayed if no corresponding doc is found.
I did that because otherwise, it forces the user to close (or to type
"u") for "uninteresting results".
(implementation is not very pretty, but it does the job).
- The mappings are only set for "python" and "man" filetypes (the
latter being useful for pydoc search inside pydoc)
- Added two mappings : <leader>pk (for pydoc -k word) and <leader>pK
(for pydoc -k WORD)